readonly_displayed_text_handler refactor (move to helpers.py)
